fix: unsafe methods not causing cache purge
Fixes bug introduced in #3733 where unsafe methods no longer make it to
the cache handler and cause a cache purge for an origin.

Also removes a duplicate test file.
