+++ lib/handler/cache-handler.js
@@ -17,11 +17,6 @@ class CacheHandler extends DecoratorHandler {
   #store
 
   /**
-   * @type {import('../../types/cache-interceptor.d.ts').default.CacheMethods}
-   */
-  #methods
-
-  /**
    * @type {import('../../types/dispatcher.d.ts').default.RequestOptions}
    */
   #requestOptions
@@ -42,14 +37,13 @@ class CacheHandler extends DecoratorHandler {
    * @param {import('../../types/dispatcher.d.ts').default.DispatchHandlers} handler
    */
   constructor (opts, requestOptions, handler) {
-    const { store, methods } = opts
+    const { store } = opts
 
     super(handler)
 
     this.#store = store
     this.#requestOptions = requestOptions
     this.#handler = handler
-    this.#methods = methods
   }
 
   /**
@@ -75,7 +69,7 @@ class CacheHandler extends DecoratorHandler {
     )
 
     if (
-      !this.#methods.includes(this.#requestOptions.method) &&
+      !util.safeHTTPMethods.includes(this.#requestOptions.method) &&
       statusCode >= 200 &&
       statusCode <= 399
     ) {
