+++ lib/interceptor/cache.js
@@ -32,7 +32,7 @@ module.exports = (opts = {}) => {
 
   return dispatch => {
     return (opts, handler) => {
-      if (!opts.origin || !methods.includes(opts.method)) {
+      if (!opts.origin || (opts.method in util.safeHTTPMethods && !methods.includes(opts.method))) {
         // Not a method we want to cache or we don't have the origin, skip
         return dispatch(opts, handler)
       }
