+++ test/interceptors/cache.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { describe, test, after } = require('node:test')
-const { strictEqual, notEqual, fail } = require('node:assert')
+const { strictEqual, notEqual, fail, equal } = require('node:assert')
 const { createServer } = require('node:http')
 const { once } = require('node:events')
 const FakeTimers = require('@sinonjs/fake-timers')
@@ -249,5 +249,42 @@ describe('Cache Interceptor', () => {
         'another-header': '123abc'
       }
     })
+  })
+
+  test('unsafe methods call the store\'s deleteByOrigin function', async () => {
+    const server = createServer((_, res) => {
+      res.end('asd')
+    }).listen(0)
+
+    after(() => server.close())
+    await once(server, 'listening')
+
+    let deleteByOriginCalled = false
+    const store = new cacheStores.MemoryCacheStore()
+
+    const originalDeleteByOrigin = store.deleteByOrigin.bind(store)
+    store.deleteByOrigin = (origin) => {
+      deleteByOriginCalled = true
+      originalDeleteByOrigin(origin)
+    }
+
+    const client = new Client(`http://localhost:${server.address().port}`)
+      .compose(interceptors.cache({ store }))
+
+    await client.request({
+      origin: 'localhost',
+      method: 'GET',
+      path: '/'
+    })
+
+    equal(deleteByOriginCalled, false)
+
+    await client.request({
+      origin: 'localhost',
+      method: 'DELETE',
+      path: '/'
+    })
+
+    equal(deleteByOriginCalled, true)
   })
 })
